+++ TeamApp/TeamApp.Infrastructure.Persistence/Entities/KanbanList.cs
@@ -13,7 +13,7 @@ namespace TeamApp.Infrastructure.Persistence.Entities
         public string KanbanListId { get; set; }
         public string KanbanListTitle{ get; set; }
         public string KanbanListBoardBelongedId { get; set; }
-        public int? KanbanListOrderInBoard { get; set; }
+        public float? KanbanListOrderInBoard { get; set; }
         public bool? KanbanListIsDeleted { get; set; }
 
         public virtual KanbanBoard KanbanBoard { get; set; }
